Created Date Time service. Added Assets path to Server info contract. Removed not needed test action for player. Added Run Skill action from player. Added skill functionaly to Entity base class. Added CSX/server scripts to loadable assets folder. Added Client Action Scripts to assets folder. Added configuration files for Combat/Skill systems.

+++ src/EventHorizon.Game.Server.Zone.Model/Gui/Options/GuiBar.cs
@@ -11,7 +11,7 @@ namespace EventHorizon.Game.Server.Zone.Model.Gui.Options
         public string BarColor { get; set; }
         public string BarDirection { get; set; }
         public string BorderColor { get; set; }
-        public int Percent { get; set; }
+        public long Percent { get; set; }
         public int BorderThickness { get; set; }
     }
 }
